refactor: Signup 페이지 컨텐츠를 app 폴더에서 pages 폴더로 이동(Next-Study-habjung#45)
